Organizzato le varie classi in più file
